feat(Gathering): submit

+++ api/entities/media-collection.js
@@ -1,9 +1,5 @@
 const mongoose = require('mongoose')
 const AWS = require('aws-sdk')
-const Media = require('./media')
-
-Media.model = mongoose.model('media', Media.fields)
-global.Media = Media
 
 let MediaCollection = {
     write: 'user',
@@ -14,22 +10,53 @@ let MediaCollection = {
     }, { timestamps: true })
 }
 
-MediaCollection.fields.post('findOneAndDelete', async function (doc, next) {
+MediaCollection.fields.pre('deleteOne', { document: true, query: false }, async function (next) {
+    const doc = this
+
+    if (doc && doc.medias) {
+        const s3 = new AWS.S3({
+            accessKeyId: process.env.S3_ID,
+            secretAccessKey: process.env.S3_SECRET
+        })
+
+        await Promise.all(doc.medias.map(async media => {
+            if (media.id) {
+                s3.deleteObject({
+                    Bucket: process.env.S3_BUCKET, Key: media.id
+                }, (e) => {
+                    console.log(e)
+                })
+            }
+
+            return true
+        }))
+    }
+
+    next()
+})
+
+MediaCollection.fields.pre('remove', async function (next) {
     const s3 = new AWS.S3({
         accessKeyId: process.env.S3_ID,
         secretAccessKey: process.env.S3_SECRET
     })
 
-    await Promise.all(doc.medias.map(async media => {
-        s3.deleteObject({
-            Bucket: process.env.S3_BUCKET, Key: media.id
-        }, (e) => {
-            console.log(e)
-        })
+    console.log(this)
 
-        return true
-    }))
-    
+    if (this.medias) {
+        await Promise.all(this.medias.map(async media => {
+            if (media.id) {
+                s3.deleteObject({
+                    Bucket: process.env.S3_BUCKET, Key: media.id
+                }, (e) => {
+                    console.log(e)
+                })
+            }
+
+            return true
+        }))
+    }
+
     next()
 })
